+++ src/components/InstitutionCard.jsx
@@ -1,26 +1,8 @@
 import React from 'react';
-import { FaTrash } from 'react-icons/fa';
 
 export default function InstitutionCard({ institution, setCurrentInstitution }) {
     function handleClick(e) {
         setCurrentInstitution(institution);
-    }
-
-    function handleSubmit() {
-        deleteInstitution();
-    }
-
-    async function deleteInstitution() {
-        try {
-            const options = {
-                method: 'DELETE'
-            };
-
-            const response = await fetch(`http://localhost:4000/institutions/${institution.ins_id}`, options);
-
-        } catch (error) {
-            console.error(error.message);
-        }
     }
 
     return (
@@ -28,11 +10,6 @@ export default function InstitutionCard({ institution, setCurrentInstitution }) 
             <img src="" alt={`${institution.name} logo`} />
             <h2>{institution.name}</h2>
             <h2>${institution.current_balance || "0.00"}</h2>
-            <form onSubmit={handleSubmit} className='delete-form institution-delete-form'>
-                <button type="submit">
-                    <FaTrash />
-                </button>
-            </form>
         </div>
     );
 }
