+++ src/components/InstitutionDetailsHeader.jsx
@@ -1,6 +1,24 @@
 import React from 'react';
+import { FaTrash } from 'react-icons/fa';
 
 export default function InstitutionDetailsHeader({ institution, accounts }) {
+    function handleSubmit() {
+        deleteInstitution();
+    }
+
+    async function deleteInstitution() {
+        try {
+            const options = {
+                method: 'DELETE'
+            };
+
+            const response = await fetch(`http://localhost:4000/institutions/${institution.ins_id}`, options);
+
+        } catch (error) {
+            console.error(error.message);
+        }
+    }
+    
     function sumBalances(resource) {
         let sum = 0;
         for (let i = 0; i < resource.length; i++) {
@@ -15,6 +33,11 @@ export default function InstitutionDetailsHeader({ institution, accounts }) {
     return (
         <div className='header details-header institution-details-header'>
             <h2 className="details-title">{`${institution.name} Details`}</h2>
+            <form onSubmit={handleSubmit} className='delete-form institution-delete-form'>
+                <button type="submit">
+                    <FaTrash />
+                </button>
+            </form>
             <table>
                 <thead>
                     <tr>
